Added book and author connection to the entities with many to many
Now the author's books can be read

+++ src/main/java/com/szelestamas/bookstorebackend/api/author/AuthorService.java
@@ -3,6 +3,8 @@ package com.szelestamas.bookstorebackend.api.author;
 import com.szelestamas.bookstorebackend.api.author.domain.Author;
 import com.szelestamas.bookstorebackend.api.author.persistence.AuthorEntity;
 import com.szelestamas.bookstorebackend.api.author.persistence.AuthorRepository;
+import com.szelestamas.bookstorebackend.core.ResourceAlreadyExistsException;
+import com.szelestamas.bookstorebackend.core.ResourceCannotBeDeletedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -29,10 +31,16 @@ public class AuthorService {
     }
 
     public Author newAuthor(Author author) {
+        if (authorRepository.findByFullName(author.getFullName()).isPresent()) {
+            throw new ResourceAlreadyExistsException(author.getFullName());
+        }
         return authorRepository.save(AuthorEntity.of(author)).toAuthor();
     }
 
     public void deleteById(long id) {
+        if (!authorRepository.findById(id).get().getBooks().isEmpty()) {
+            throw new ResourceCannotBeDeletedException("Author still have books, cannot delete it.");
+        }
         authorRepository.deleteById(id);
     }
 
@@ -40,7 +48,7 @@ public class AuthorService {
         Optional<AuthorEntity> authorEntity = authorRepository.findById(id);
         if (authorEntity.isPresent()) {
             AuthorEntity authorEntityToUpdate = authorEntity.get();
-            authorEntityToUpdate.setFullName(author.fullName());
+            authorEntityToUpdate.setFullName(author.getFullName());
             return authorRepository.save(authorEntityToUpdate).toAuthor();
         } else {
             throw new NoSuchElementException("Author not found");
